+++ serde_with/CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 * Add support for `Range`, `RangeFrom`, `RangeTo`, `RangeInclusive` (#851)
     `RangeToInclusive` is currently unsupported by serde.
 * Add `schemars` implementations for `Bound`, `Range`, `RangeFrom`, `RangeTo`, `RangeInclusive`.
+* Added support for `schemars` v1 under the `schemars_1` feature flag
 
 ## [3.13.0] - 2025-06-14